feat: 최근 사진 업로드 시간 조회 기능 Photo API 로 이동
