+++ src/main/java/com/pohyoja/picchargeserver/domain/family/controller/FamilyController.java
@@ -5,7 +5,6 @@ import com.pohyoja.picchargeserver.common.security.JwtUserDetails;
 import com.pohyoja.picchargeserver.domain.family.dto.response.FamilyResponse;
 import com.pohyoja.picchargeserver.domain.family.dto.response.FamilyUserNamesResponse;
 import com.pohyoja.picchargeserver.domain.family.dto.response.InviteCodeResponse;
-import com.pohyoja.picchargeserver.domain.family.dto.response.LatestUploadTimeResponse;
 import com.pohyoja.picchargeserver.domain.family.service.FamilyService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -112,27 +111,6 @@ public class FamilyController {
         log.info("User {} leaving family {}", userDetails.uid(), familyId);
         familyService.leaveFamily(familyId, userDetails.uid());
         return BaseResponse.onSuccess(null);
-    }
-
-    @GetMapping("/families/{familyId}/photos/latest-upload-time")
-    @Operation(
-            summary = "가족에 가장 최근에 업로드된 사진의 업로드 시간 가져오기",
-            description = "가족 방에 가장 최근에 업로드된 사진의 시간을 조회합니다. 가족 구성원만 접근 가능합니다."
-    )
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "조회 성공"),
-            @ApiResponse(responseCode = "404", description = "가족을 찾을 수 없음", content = @Content),
-            @ApiResponse(responseCode = "403", description = "가족 구성원이 아님", content = @Content)
-    })
-    public BaseResponse<LatestUploadTimeResponse> getLatestPhotoUploadTime(
-            @Parameter(description = "조회할 가족 ID", required = true, example = "1")
-            @PathVariable Long familyId,
-
-            @Parameter(description = "현재 인증된 사용자 정보", hidden = true)
-            @AuthenticationPrincipal JwtUserDetails userDetails) {
-
-        log.info("Getting latest photo upload time for family ID: {}", familyId);
-        return BaseResponse.onSuccess(familyService.getLatestPhotoUploadTime(familyId, userDetails.uid()));
     }
 
     @PostMapping("/families/{familyId}/invite-codes")
